backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Validation
email-validator>=2.0.0
